+++ +gams/+transfer/+alias/Set.m
@@ -258,7 +258,37 @@ classdef Set < gams.transfer.alias.Abstract
         end
 
         function flag = isValid(obj, varargin)
-            flag = false;
+
+            verbose = 0;
+            if nargin > 1
+                verbose = max(0, min(2, varargin{1}));
+            end
+
+            if ~obj.container_.hasSymbols(obj.name_) || obj.container_.getSymbols(obj.name_) ~= obj
+                msg = 'Alias is not contained in its linked container.';
+                switch verbose
+                case 1
+                    warning(msg);
+                case 2
+                    error(msg);
+                end
+                flag = false;
+                return
+            end
+
+            if ~obj.alias_with_.isValid()
+                msg = 'Linked symbol is invalid.';
+                switch verbose
+                case 1
+                    warning(msg);
+                case 2
+                    error(msg);
+                end
+                flag = false;
+                return
+            end
+
+            flag = true;
         end
 
         function dom_violations = getDomainViolations(obj)